[IMP]Improved Code
