+++ hotel_housekeeping/models/__init__.py
@@ -1,3 +1,6 @@
 # See LICENSE file for full copyright and licensing details.
 
 from . import hotel_housekeeping
+from . import hotel_activity
+from . import hotel_housekeeping_activities
+from . import hotel_housekeeping_activity_type
