+++ hotel_housekeeping/models/hotel_housekeeping.py
@@ -1,76 +1,7 @@
 # See LICENSE file for full copyright and licensing details.
 
-import time
-from odoo.tools import DEFAULT_SERVER_DATE_FORMAT
 from odoo import api, fields, models, _
-from odoo.osv import expression
 from odoo.exceptions import ValidationError
-
-
-class HotelHousekeepingActivityType(models.Model):
-
-    _name = 'hotel.housekeeping.activity.type'
-    _description = 'Activity Type'
-
-    name = fields.Char('Name', required=True)
-    activity_id = fields.Many2one('hotel.housekeeping.activity.type',
-                                  'Activity Type')
-
-    @api.multi
-    def name_get(self):
-        def get_names(cat):
-            """ Return the list [cat.name, cat.activity_id.name, ...] """
-            res = []
-            while cat:
-                res.append(cat.name)
-                cat = cat.activity_id
-            return res
-        return [(cat.id, " / ".join(reversed(get_names(cat)))) for cat in self]
-
-    @api.model
-    def name_search(self, name, args=None, operator='ilike', limit=100):
-        if not args:
-            args = []
-        if name:
-            # Be sure name_search is symmetric to name_get
-            category_names = name.split(' / ')
-            parents = list(category_names)
-            child = parents.pop()
-            domain = [('name', operator, child)]
-            if parents:
-                names_ids = self.name_search(' / '.join(parents), args=args,
-                                             operator='ilike', limit=limit)
-                category_ids = [name_id[0] for name_id in names_ids]
-                if operator in expression.NEGATIVE_TERM_OPERATORS:
-                    categories = self.search([('id', 'not in', category_ids)])
-                    domain = expression.OR([[('activity_id', 'in',
-                                              categories.ids)], domain])
-                else:
-                    domain = expression.AND([[('activity_id', 'in',
-                                               category_ids)], domain])
-                for num in range(1, len(category_names)):
-                    domain = [[('name', operator,
-                                ' / '.join(category_names[-1 - num:]))], domain]
-                    if operator in expression.NEGATIVE_TERM_OPERATORS:
-                        domain = expression.AND(domain)
-                    else:
-                        domain = expression.OR(domain)
-            categories = self.search(expression.AND([domain, args]),
-                                     limit=limit)
-        else:
-            categories = self.search(args, limit=limit)
-        return categories.name_get()
-
-
-class HotelActivity(models.Model):
-
-    _name = 'hotel.activity'
-    _description = 'Housekeeping Activity'
-
-    h_id = fields.Many2one('product.product', 'Product', required=True,
-                           delegate=True, ondelete='cascade', index=True)
-    categ_id = fields.Many2one('hotel.housekeeping.activity.type',
-                               string='Category')
 
 
 class HotelHousekeeping(models.Model):
@@ -179,54 +110,3 @@ class HotelHousekeeping(models.Model):
             'clean': True,
             'dirty': False,
         })
-
-
-class HotelHousekeepingActivities(models.Model):
-
-    _name = "hotel.housekeeping.activities"
-    _description = "Housekeeping Activities "
-
-    a_list = fields.Many2one('hotel.housekeeping', string='Reservation')
-    today_date = fields.Date('Today Date')
-    activity_name = fields.Many2one('hotel.activity',
-                                    string='Housekeeping Activity')
-    housekeeper_id = fields.Many2one('res.users', string='Housekeeper',
-                                  required=True)
-    clean_start_time = fields.Datetime('Clean Start Time',
-                                       required=True)
-    clean_end_time = fields.Datetime('Clean End Time', required=True)
-    dirty = fields.Boolean('Dirty',
-                           help='Checked if the housekeeping activity'
-                           'results as Dirty.')
-    clean = fields.Boolean('Clean', help='Checked if the housekeeping'
-                           'activity results as Clean.')
-
-    @api.constrains('clean_start_time', 'clean_end_time')
-    def _check_clean_start_time(self):
-        '''
-        This method is used to validate the clean_start_time and
-        clean_end_time.
-        ---------------------------------------------------------
-        @param self: object pointer
-        @return: raise warning depending on the validation
-        '''
-        if self.clean_start_time >= self.clean_end_time:
-            raise ValidationError(_(
-                'Start Date Should be less than the End Date!'))
-
-    @api.model
-    def default_get(self, fields):
-        """
-        To get default values for the object.
-        @param self: The object pointer.
-        @param fields: List of fields for which we want default values
-        @return: A dictionary which of fields with values.
-        """
-        if self._context is None:
-            self._context = {}
-        res = super().default_get(fields)
-        if self._context.get('room_id', False):
-            res.update({'room_id': self._context['room_id']})
-        if self._context.get('today_date', False):
-            res.update({'today_date': self._context['today_date']})
-        return res
